+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Change Log
 * Seems no one does use PyPI for installation, .pyui files are now included
 * Comment line with ``# `` instead of just ``#`` (#12)
 * ``Cmd U`` to run unit tests (experimental)
+* ``Ctrl Tab`` (or ``Cmd Shift ]``) selects next tab
+* ``Ctrl Shift Tab`` (or ``Cmd Shift [``) selects previous tab
+* ``Cmd 1..9`` selects specific tab
 
 
 ``0.0.13`` (2017-08-24)